[refactor] 게시글 수정 페이지 클라이언트 분리
